requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
psycopg2==2.9.3
asyncpg==2.9.9
PyJWT == 1.7.1
python-decouple==3.3